Shark chase: one copy of the rules
The two variant folders held the same `gameplay.ts` twice, differing
only in `MAX_TURN` and `startBoard` — the moves, `isGameEnd` and
`getWinnerIndex` were identical text.

`makeGameplay(maxTurn)` in the shared `gameplay.ts` takes them, as
`makeGeometry(size)` already takes the lake's size. A variant states the
two things that are its own and takes the rest:

    export const { isGameEnd, getWinnerIndex, moves } = makeGameplay(MAX_TURN);

To check it: diff the two old files against each other — `11`/`15` and
the start board are the whole of it — then read the extraction as the
4 × 4 copy with `11` replaced by the parameter.

The bots are untouched, and every existing spec imports the same names
from the same places as before.
